Refactor: lecture 도메인으로 registartion 도메인 이동
