+++ src/test/java/com/hhplus/special_lecture_service/integrationTest/LectureLimitThirtyIntegrationTest.java
@@ -1,0 +1,20 @@
+package com.hhplus.special_lecture_service.integrationTest;
+
+import com.hhplus.special_lecture_service.integrationTest.setUp.LectureSetUp;
+import com.hhplus.special_lecture_service.integrationTest.setUp.UserSetUp;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class LectureLimitThirtyIntegrationTest extends BaseIntegrationTest{
+
+    @Autowired
+    private UserSetUp userSetUp;
+
+    @Autowired
+    private LectureSetUp lectureSetUp;
+
+
+    /*@Test
+    @DisplayName("")*/
+}
