+++ src/test/java/com/hhplus/special_lecture_service/integrationTest/setUp/RegistrationSetUp.java
@@ -2,13 +2,12 @@ package com.hhplus.special_lecture_service.integrationTest.setUp;
 
 import com.hhplus.special_lecture_service.domain.common.StatusType;
 import com.hhplus.special_lecture_service.domain.lecture.Lecture;
-import com.hhplus.special_lecture_service.domain.registration.Registration;
+import com.hhplus.special_lecture_service.domain.lecture.Registration;
 import com.hhplus.special_lecture_service.domain.user.User;
 import com.hhplus.special_lecture_service.infrastructure.registration.RegistrationJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.sql.Timestamp;
 import java.time.LocalDate;
 import java.time.LocalTime;
