+++ src/test/java/com/hhplus/special_lecture_service/unitTest/model/RegistrationTest.java
@@ -1,9 +1,8 @@
 package com.hhplus.special_lecture_service.unitTest.model;
 
 import com.hhplus.special_lecture_service.common.exception.OverCapacityException;
-import com.hhplus.special_lecture_service.domain.common.StatusType;
 import com.hhplus.special_lecture_service.domain.lecture.Lecture;
-import com.hhplus.special_lecture_service.domain.registration.Registration;
+import com.hhplus.special_lecture_service.domain.lecture.Registration;
 import com.hhplus.special_lecture_service.domain.user.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -11,11 +10,7 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.sql.Date;
-import java.sql.Time;
-import java.sql.Timestamp;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
